Separated 2D- and 3D navigation
