+++ source/navigationmodifier2D.ts
@@ -1,0 +1,10 @@
+import {AbstractNavigationModifier} from "./abstractnavigationmodifier";
+
+export class Navigationmodifier2D extends AbstractNavigationModifier {
+    rotate(): void {
+    }
+
+    scale(step?: number): void {
+    }
+
+}
